fix: change AlpacaBar.Volume to double for crypto decimal volumes
Alpaca crypto volumes are fractional (e.g. 0.5 BTC), not integers.

Made-with: Cursor

+++ services/workers/data-fetcher-2.0/src/DataFetcher.Worker/Application/Providers/Alpaca/AlpacaCryptoBackfillService.cs
@@ -82,7 +82,7 @@ public class AlpacaCryptoBackfillService : IAlpacaCryptoBackfillService
                     HighPrice = (decimal)bar.High,
                     LowPrice = (decimal)bar.Low,
                     ClosePrice = (decimal)bar.Close,
-                    Volume = bar.Volume,
+                    Volume = (long)bar.Volume,
                     MarketCap = null
                 }).ToList();
