+++ services/workers/data-fetcher-2.0/src/DataFetcher.Worker/Application/Providers/Alpaca/AlpacaCryptoFetchService.cs
@@ -83,7 +83,7 @@ public class AlpacaCryptoFetchService : IAlpacaCryptoFetchService
                     HighPrice = (decimal)bar.High,
                     LowPrice = (decimal)bar.Low,
                     ClosePrice = (decimal)bar.Close,
-                    Volume = bar.Volume,
+                    Volume = (long)bar.Volume,
                     MarketCap = null
                 }).ToList();
